+++ src/test/java/com/github/clave/keycloak/ClaveAttributeMapperTest.java
@@ -12,4 +12,12 @@ public class ClaveAttributeMapperTest {
         assertEquals("Attribute Importer", mapper.getDisplayCategory());
         assertArrayEquals(new String[]{ClaveIdentityProviderFactory.PROVIDER_ID}, mapper.getCompatibleProviders());
     }
+
+    @Test
+    public void testConstants() {
+        assertEquals("http://eidas.europa.eu/attributes/naturalperson/PersonIdentifier", ClaveAttributeMapper.PERSON_IDENTIFIER);
+        assertEquals("http://eidas.europa.eu/attributes/naturalperson/CurrentGivenName", ClaveAttributeMapper.GIVEN_NAME);
+        assertEquals("http://eidas.europa.eu/attributes/naturalperson/CurrentFamilyName", ClaveAttributeMapper.FAMILY_NAME);
+        assertEquals("http://eidas.europa.eu/attributes/naturalperson/DateOfBirth", ClaveAttributeMapper.DATE_OF_BIRTH);
+    }
 }